Update "Set up JDBC Driver" section
The "### Set up JDBC Driver for SQL Server database" section is outdated in 2019.1 branch.
The link from step 1 (Download the Microsoft JDBC driver v6.0+) used to point to version 6.0, but now it points to version 6.2 on Microsoft site. This causes problem with step 3 where it says:
In step 3 it says "Copy the sqljdbc42.jar" and is 6.2 the name of the jar is different. There is actually two of them: 
mssql-jdbc-6.2.2.jre7.jar
mssql-jdbc-6.2.2.jre8.jar

I changed the Microsoft JDBC driver link to point to the page with different driver versions to let the user pick. In step 3 I added the correction for the new jar file naming inside the downloaded package.

Please review/change as necessary.