Changed 'buildtype-graphs' type using lower case t
Changed 'buildtype-graphs' type using lower case t as mentioned in the xml example. I was cracking my head around this quite some time till I realized that the type in the actual xml needs to be lower case (I initially copied it from the text and not the xml).  To avoid this potential confusion for others I propose to change the case in the description as well.