Update integrating-teamcity-with-vcs-hosting-services.md
Changes for:
- consistency between sections
- readability (long sentences, fragments, or sentences with more than one idea in them)
- minor grammar issues (articles)
- style (simpler global English, reduced ambiguity)
- structure (logical scannable steps, breaking down steps)

I would like to change formatting for "Click the _register application_ link" to use bold instead of italics. I didn't change it because I saw that the rest of the documentation uses a combination of bold/italics for GUI elements. I assume the style guide has a rule for when to use italics, so I would check the rule before making changes. However, bold is much easier to read, especially when readers are scanning steps.