Update AND operator support info
Since the current TC versions use a new scheme in indexes, the `AND` clause may not yield the same results as it used to and its use is recommended. Instead, the user should simply separate the search criteria with white space (an equivalent of the `OR` clause). The topmost search results should be the most relevant, i.e. ones that satisfy all criteria.